Add another mode and bump to V2.0
There was an unintended behaviour in the gesture which was that if you tap on other corner and then tap on other corner while holding previous one, the screen would still sleep because the code only checks for detection trigger at one corner and power trigger at other corner. This commit fixes it and adds it as optional by sysfs.

The value '2' of "t2s_switch" will enable the sleeping of screen even when you touch both corners.
Earlier this was unintended so edited code to make it a separate behaviour.
Default value is '1' which will check whether the finger goes through the middle part of path which will set the checkpoint variable to 1.
When we reach the other corner, it will check value of checkpoint to sleep.

If we don't give input to middle part of arc (as visible from code), t2s will reset and break the detection until next time a corner tap is detected.

Of course the mode 1 can also be cheated with 3 fingers (two at both corners, and one in middle) but that prevents accidental wakeup.
Mode '2' is definitely faster so I thought of leaving it up to the user to change it using sysfs.

Signed-off-by: dopaemon <polarisdp@gmail.com>