scsi: ufs: set IRQ's affinity to CPU1
CPU0 handles most of the IRQs and UFS IRQ is one of the most
frequent ones, however it doesn't directly affect UX so it's
not worth it to move it to perf cluster (we only have 2 big cores).

Let's move it to CPU1 to reduce the load on CPU0 a little.

Signed-off-by: Yaroslav Furman <yaro330@gmail.com>
Signed-off-by: Oktapra Amtono <oktapra.amtono@gmail.com>